tablo mesajlar ve düzzenlemeler

+++ src/app/features/ihale/ihale-olustur/kesif-olustur/kesif-olustur.component.ts
@@ -1,8 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { MenuItem, Message, TreeNode } from 'primeng/api';
+import { MenuItem, Message, MessageService, TreeNode } from 'primeng/api';
 import { TablodataService } from '../../../../services/tablodata.service';
 import { Column } from '../../../../models/column.interface';
-import { MessagesModule } from 'primeng/messages';
 import { ToolbarModule } from 'primeng/toolbar';
 import { MenuModule } from 'primeng/menu';
 import { MultiSelectModule } from 'primeng/multiselect';
@@ -16,6 +15,7 @@ import { DialogModule } from 'primeng/dialog';
 import { NgClass } from '@angular/common';
 import { RouterLink } from '@angular/router';
 import { IhaleOlusturComponent } from '../ihale-olustur.component';
+import { ToastModule } from 'primeng/toast';
 
 import { Subscription } from 'rxjs';
 import { XlsxService } from '../../../../services/xlsx.service';
@@ -25,8 +25,8 @@ import { XlsxService } from '../../../../services/xlsx.service';
 @Component({
   selector: 'app-kesif-olustur',
   standalone: true,
-  imports: [MessagesModule, MenuModule, ToolbarModule, MultiSelectModule, 
-    FormsModule, ButtonModule, TreeTableModule, InputNumberModule,
+  imports: [ MenuModule, ToolbarModule, MultiSelectModule, 
+    FormsModule, ButtonModule, TreeTableModule, InputNumberModule, ToastModule,
    ContextMenuModule, NgClass, InputTextModule, DialogModule, RouterLink, IhaleOlusturComponent],
   templateUrl: './kesif-olustur.component.html',
   styleUrl: './kesif-olustur.component.scss'
@@ -59,7 +59,7 @@ delCols: Column[] =[];
         this.visible = true;
     }
 
-  constructor(private dataService: TablodataService, private excelService: XlsxService) {}
+  constructor(private dataService: TablodataService, private excelService: XlsxService, private messageService: MessageService) {}
 
   ngOnInit(): void {
     // verileri yükle
@@ -89,11 +89,7 @@ delCols: Column[] =[];
     { label: 'Diğer Sütun', icon: 'pi pi-server', command: (event) => this.showDialog() }
     ];
 
-
-    //mesajları oluştur
-    this.messages = [
-      { severity: 'info', summary: 'Keşif Oluşturma', detail: 'Tablo üzerinde değişiklik yapmak için sağ tuşa tıklayarak açılan menüyü kullanabilirsiniz.' },
-    ];
+   
   }
 
   addRowToNode(node: TreeNode) {
@@ -108,37 +104,74 @@ delCols: Column[] =[];
     this.expandAllNodes(this.files);
   }
 
+  showInfo() {
+    this.messageService.add({ severity: 'info', summary: 'Bilgi', detail: 'Tablo üzerinde değişiklik yapmak için sağ tuşa tıklayarak açılan menüyü kullanabilirsiniz.' });
+  }
   deleteNode(selectedNode: TreeNode<any>): void {
 
-    //Satırı sil ve poz noları güncelle
-    this.dataService.deleteRow(selectedNode,this.files);
-
-    //Toplamı güncelle
-    if(selectedNode.parent) {
-      if (selectedNode.parent.children) {
-        let allToplam = 0;
-        for(const child of selectedNode.parent.children) {
-          allToplam += child.data.Toplam
-          
+    if(selectedNode.children) {
+      if(selectedNode.children.length>0) {
+        if(!window.confirm('Seçtiğiniz başlığın tüm alt kırılımlarıyla birlikte silinmesini onaylıyor musunuz?')) {
+          this.messageService.add({ severity: 'error', summary: 'İptal', detail: 
+          'Satır silinmesi işlemi kullanıcı tarafından iptal edildi.' });
+ 
+          return;
         }
-        selectedNode.parent.data.Toplam=allToplam;
       }
-    }
+    } 
+    //Satırı sil ve poz noları güncelle
+    const result=this.dataService.deleteRow(selectedNode,this.files);
+    if(result) {
+        //Toplamı güncelle
+        if(selectedNode.parent) {
+          if (selectedNode.parent.children) {
+            let allToplam = 0;
+            for(const child of selectedNode.parent.children) {
+              allToplam += child.data.Toplam
+              
+            }
+            selectedNode.parent.data.Toplam=allToplam;
+          }
+        }
+        this.messageService.add({ severity: 'success', summary: 'Satır Silindi', detail: 
+        'Satır silinmesi işlemi başarıyla gerçekleştirildi.' });
+        this.updateView();
+        return;
+      }
+      else {
+        this.messageService.add({ severity: 'error', summary: 'Satır Silinemedi', detail: 
+        'Tabloda başka ana başlık olmadığından seçtiğiniz ana başlık silinemedi.' });
+        return;
+      }
     
+    
+ 
 
-    this.updateView();
+   
   }
  
   deleteCol(columns: Column[]) {
     for (let i=0; i<columns.length; i++){
       const index = this.cols.indexOf(columns[i]);
       if(i>-1){
+        this.messageService.add({ severity: 'success', summary: 'Sütunlar Silindi', detail: 
+          'Seçtiğiniz sütun "' + this.cols[index].header +'" başarıyla silindi.' });
         this.cols.splice(index,1)
+        
       }
     }
+    
+ 
     this.delCols =[];   
   }
   addOtherCol(name: string) {
+    if(this.cols.find(x=>x.header==name)) {
+      console.log('exist')
+      this.messageService.add({ severity: 'error', summary: 'Sütun Mevcut', detail: 'Eklemek İstediğiniz sütunla aynı isimde bir sütun zaten olduğu için ekleme yapılmadı.' });
+      return;
+    }
+
+    
     //Sütunları Güncelle
     
     this.dataService.addOtherCol(name, this.cols)
@@ -153,10 +186,7 @@ delCols: Column[] =[];
   addBirimCol(name: string = 'Diğer') {
     if(this.cols.find(x=>x.header==name+' Birim Fiyat')) {
       console.log('exist')
-      this.messages.push(    
-          { severity: 'error', summary: 'Sütun zaten mevcut', detail: 
-          'Eklemek istediğiniz sütunla aynı isimde sütun tabloda bulunduğu için değişiklik yapılmadı.' },
-      );
+      this.messageService.add({ severity: 'error', summary: 'Sütun Mevcut', detail: 'Eklemek İstediğiniz sütunla aynı isimde bir sütun zaten olduğu için ekleme yapılmadı.' });
       return;
     }
